contact_app3/requirements.txt
click==8.3.0
colorama==0.4.6
python-dotenv==1.1.1
